Update bug that splits shapes when hit by volatile

+++ src/conductor/controller/state.rs
@@ -479,13 +479,66 @@ impl State {
               and    _o.shape <> ?1
               and    (o.x,o.y) = (_o.x,_o.y))
             "#, params![shape], |row| row.get(0))? {
-              // Remove all objects at position the position the Volatile is moved to
-              tx.execute(r#"
-                delete from objects
-                where  (x,y) = (?1,?2)
-              "#, params![there_x,there_y])?;
-              return Ok(Some((here_shape, vec![Object::new(0, 0, 0, "Removed".to_string(), there)], None)));
-            }
+              let o_shape = tx.query_row(r#"
+                select o.shape from objects as o where (o.x,o.y) = (?1,?2) and o.shape <> ?3
+              "#, params![there_x, there_y, shape], |row| row.get(0))?;
+              if let Some(shape) = o_shape {
+                // Keep the completed shape positions
+                tx.execute("create temporary table completed_shape as select * from objects as o where o.shape = ?1", params![shape])?;
+                // Remove all objects at position the position the Volatile is moved to
+                tx.execute(r#"
+                  delete from objects
+                  where  (x,y) = (?1,?2)
+                "#, params![there_x,there_y])?;
+                // Update shapes
+                while {
+                  let new_shape: i32 = tx.query_row("select nextval('shape_seq_id')", params![], |row| row.get(0))?;
+                  tx.execute(r#"
+                  update objects
+                  set    shape = ?2
+                  where  (x,y) in (
+                    with recursive one_shape(shape,connectors,kind,x,y) as (
+                      (select o.shape, o.connectors, o.kind, o.x, o.y
+                      from   objects as o
+                      where  shape = ?1
+                      limit  1)
+                        union
+                      select o.shape, o.connectors, o.kind, o.x, o.y
+                      from   one_shape as os, objects as o
+                      where  os.shape = o.shape
+                      and    "connects?"(os.connectors, os.kind, os.x, os.y, o.connectors, o.kind, o.x, o.y)
+                    )
+                    select (os.x,os.y)
+                    from   one_shape as os
+                  )
+                "#, params![shape, new_shape])? > 0 } { }
+                // Query all objects that have previously been `shape` and set
+                let mut selected_shape = Some(shape);
+                let there_shape =
+                  State::query_objects_via_statement(
+                    tx.prepare(r#"
+                      select cs.id, o.shape, o.connectors, o.kind::text, cs.x, cs.y
+                      from   completed_shape as cs left join objects as o on (cs.x,cs.y) = (o.x,o.y);
+                    "#)?, params![],
+                    |row| {
+                      let id = row.get(0)?;
+                      let o_shape: Option<i32> = row.get(1)?;
+                      let pos = (row.get(4)?,row.get(5)?);
+                      if pos == there { selected_shape = o_shape }
+                      if let Some(shape) = o_shape {
+                        Ok(Object::new_with_color(id, shape, row.get(2)?, row.get(3)?, pos, Some(Color::DarkGrey)))
+                      } else {
+                        Ok(Object::new(id, 0, 0, "Removed".to_string(), pos))
+                      }
+                    }
+                  )?;
+                  // Clean up
+                  tx.execute("drop table completed_shape", params![])?;
+                return Ok(Some((here_shape,there_shape,selected_shape)));
+              } else {
+                return Ok(None);
+              }
+          }
           // Merge shapes, if any are adjacent and then do things
           // If number of merged objects is larger than 0, check for doors
           if tx.execute(r#"
@@ -525,6 +578,7 @@ impl State {
                   where shape = ?1
                   and   kind = 'Door'
               "#, params![shape])?;
+
               // Update shapes
               while {
                 let new_shape: i32 = tx.query_row("select nextval('shape_seq_id')", params![], |row| row.get(0))?;
